Authentication and notes-user relation

+++ NotesAppAPI/Services/NotesService.cs
@@ -16,8 +16,9 @@ namespace NotesAppAPI.Services
             connectionString = configuration.GetConnectionString("NotesAppConnectionString")!;
         }
 
-        public async Task<dtoResponse<List<dtoNote>>> GetNotes()
+        public async Task<dtoResponse<List<dtoNote>>> GetNotes(int userId)
         {
+            System.Console.WriteLine(userId);
             dtoResponse<List<dtoNote>> response = new dtoResponse<List<dtoNote>>();
             List<dtoNote> notes = new List<dtoNote>();
 
@@ -27,7 +28,7 @@ namespace NotesAppAPI.Services
                 {
                     conn.Open();
 
-                    string query = "SELECT * FROM notes ORDER BY 5 DESC, 4 DESC";
+                    string query = string.Format("SELECT id, description, read, created_at, last_updated FROM notes WHERE user_id = {0} ORDER BY 3 DESC, 5 DESC, 4 DESC", userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     NpgsqlDataReader reader = await cmd.ExecuteReaderAsync();
@@ -58,8 +59,7 @@ namespace NotesAppAPI.Services
             return response;
         }
 
-
-        public async Task<dtoResponse<bool>> AddNote(dtoAddNote newNote)
+        public async Task<dtoResponse<bool>> AddNote(dtoAddNote newNote, int userId)
         {
             dtoResponse<bool> response = new dtoResponse<bool>();
             bool inserted = false;
@@ -71,8 +71,8 @@ namespace NotesAppAPI.Services
                     conn.Open();
 
                     string query = string.Format(
-                            @"INSERT INTO notes(description, read, created_at, last_updated)
-                            VALUES('{0}', DEFAULT, DEFAULT, DEFAULT)", newNote.description);
+                            @"INSERT INTO notes(description, read, created_at, last_updated, user_id)
+                            VALUES('{0}', DEFAULT, DEFAULT, DEFAULT, {1})", newNote.description, userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     inserted = await cmd.ExecuteNonQueryAsync() == 1;
@@ -94,7 +94,7 @@ namespace NotesAppAPI.Services
             return response;
         }
 
-        public async Task<dtoResponse<bool>> DeleteNote(int id)
+        public async Task<dtoResponse<bool>> DeleteNote(int id, int userId)
         {
             dtoResponse<bool> response = new dtoResponse<bool>();
             bool deleted = false;
@@ -105,7 +105,7 @@ namespace NotesAppAPI.Services
                 {
                     conn.Open();
 
-                    string query = string.Format("DELETE FROM notes WHERE id = {0}", id);
+                    string query = string.Format("DELETE FROM notes WHERE id = {0} AND user_id = {1}", id, userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     deleted = await cmd.ExecuteNonQueryAsync() == 1;
@@ -127,7 +127,7 @@ namespace NotesAppAPI.Services
             return response;
         }
 
-        public async Task<dtoResponse<dtoNote>> GetNoteById(int id)
+        public async Task<dtoResponse<dtoNote>> GetNoteById(int id, int userId)
         {
             dtoResponse<dtoNote> response = new dtoResponse<dtoNote>();
             dtoNote? result = null;
@@ -138,7 +138,7 @@ namespace NotesAppAPI.Services
                 {
                     conn.Open();
 
-                    string query = "SELECT * FROM notes WHERE id=" + id;
+                    string query = string.Format("SELECT id, description, read, created_at, last_updated FROM notes WHERE id = {0} AND user_id = {1}", id, userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     NpgsqlDataReader reader = await cmd.ExecuteReaderAsync();
@@ -149,13 +149,16 @@ namespace NotesAppAPI.Services
                         {
                             id = (int)reader["id"],
                             description = (string)reader["description"],
-                            read = (bool)reader["read"]
+                            read = (bool)reader["read"],
+                            createdAt = (DateTime)reader["created_at"],
+                            lastUpdated = (DateTime)reader["last_updated"]
                         };
                     }
 
                     cmd.Dispose();
                     conn.Close();
                 }
+                response.data = result;
 
                 if (result == null)
                     Utils.SetResponseError<dtoNote>(response, dtoResponseMessageCodes.NOT_EXISTS);
@@ -168,7 +171,7 @@ namespace NotesAppAPI.Services
             return response;
         }
 
-        public async Task<dtoResponse<bool>> UpdateNote(dtoUpdateNote modifiedNote)
+        public async Task<dtoResponse<bool>> UpdateNote(dtoUpdateNote modifiedNote, int userId)
         {
             dtoResponse<bool> response = new dtoResponse<bool>();
             bool updated = false;
@@ -179,7 +182,9 @@ namespace NotesAppAPI.Services
                 {
                     conn.Open();
 
-                    string query = string.Format("UPDATE notes SET read=false, last_updated=DEFAULT, description='{1}' WHERE id={0}", modifiedNote.id, modifiedNote.description);
+                    string query = string.Format(
+                            @"UPDATE notes SET read=false, last_updated=DEFAULT, description='{0}'
+                            WHERE id = {1} AND user_id = {2}", modifiedNote.description, modifiedNote.id, userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     updated = await cmd.ExecuteNonQueryAsync() == 1;
@@ -187,6 +192,8 @@ namespace NotesAppAPI.Services
                     cmd.Dispose();
                     conn.Close();
                 }
+
+                response.data = updated;
 
                 if (!updated)
                     Utils.SetResponseOperationNotPerformedError<bool>(response);
@@ -199,7 +206,7 @@ namespace NotesAppAPI.Services
             return response;
         }
 
-        public async Task<dtoResponse<bool>> MarkAsRead(int id)
+        public async Task<dtoResponse<bool>> MarkAsRead(int id, int userId)
         {
             dtoResponse<bool> response = new dtoResponse<bool>();
             bool read = false;
@@ -210,7 +217,7 @@ namespace NotesAppAPI.Services
                 {
                     conn.Open();
 
-                    string query = string.Format("UPDATE notes SET read=true WHERE id={0}", id);
+                    string query = string.Format("UPDATE notes SET read = true WHERE id = {0} AND user_id = {1}", id, userId);
 
                     NpgsqlCommand cmd = new NpgsqlCommand(query, conn);
                     read = await cmd.ExecuteNonQueryAsync() == 1;
@@ -218,6 +225,8 @@ namespace NotesAppAPI.Services
                     cmd.Dispose();
                     conn.Close();
                 }
+
+                response.data = read;
 
                 if (!read)
                     Utils.SetResponseError<bool>(response, dtoResponseMessageCodes.OPERATION_NOT_PERFORMED);
